vm: Move host module to the root of the lib
