+++ vm/src/cache.rs
